+++ src/main/java/org/truffleruby/core/thread/ThreadNodes.java
@@ -606,20 +606,29 @@ public abstract class ThreadNodes {
     }
 
     @Primitive(name = "thread_get_exception")
-    public static abstract class GetThreadLocalExceptionNode extends PrimitiveArrayArgumentsNode {
+    public static abstract class ThreadGetExceptionNode extends PrimitiveArrayArgumentsNode {
 
         @Specialization
         public DynamicObject getException(VirtualFrame frame,
                 @Cached("create()") GetCurrentRubyThreadNode getThreadNode) {
-            return Layouts.THREAD.getThreadLocalGlobals(getThreadNode.executeGetRubyThread(frame)).exception;
+            return getLastException(getThreadNode.executeGetRubyThread(frame));
         }
+
+        private static DynamicObject getLastException(DynamicObject currentThread) {
+            return Layouts.THREAD.getThreadLocalGlobals(currentThread).exception;
+        }
+
+        public static DynamicObject getLastException(RubyContext context) {
+            return getLastException(context.getThreadManager().getCurrentThread());
+        }
+
     }
 
     @Primitive(name = "thread_get_return_code")
-    public static abstract class GetThreadLocalReturnCodeNode extends PrimitiveArrayArgumentsNode {
+    public static abstract class ThreadGetReturnCodeNode extends PrimitiveArrayArgumentsNode {
 
         @Specialization
-        public DynamicObject getException(VirtualFrame frame,
+        public DynamicObject getExitCode(VirtualFrame frame,
                 @Cached("create()") GetCurrentRubyThreadNode getThreadNode) {
             return Layouts.THREAD.getThreadLocalGlobals(getThreadNode.executeGetRubyThread(frame)).processStatus;
         }